src/main.py
# src/main.py

import sys

roman_numerals = {
    'I': 1, 'V': 5, 'X': 10, 'L': 50, 'C': 100, 'D': 500, 'M': 1000
}

def roman_to_int(s):
    """
    Convert a Roman numeral to an integer.

    Args:
        s (str): The Roman numeral string.

    Returns:
        int: The integer representation of the Roman numeral.
    """
    total = 0
    prev_value = 0
    for char in reversed(s):
        value = roman_numerals[char]
        if value < prev_value:
            total -= value
        else:
            total += value
        prev_value = value
    return total

def int_to_roman(num):
    """
    Convert an integer to a Roman numeral.

    Args:
        num (int): The integer to convert.

    Returns:
        str: The Roman numeral representation of the integer.
    """
    val = [
        1000, 900, 500, 400,
        100, 90, 50, 40,
        10, 9, 5, 4,
        1
    ]
    syb = [
        "M", "CM", "D", "CD",
        "C", "XC", "L", "XL",
        "X", "IX", "V", "IV",
        "I"
    ]
    roman_num = ''
    i = 0
    while num > 0:
        for _ in range(num // val[i]):
            roman_num += syb[i]
            num -= val[i]
        i += 1
    return roman_num

def calculate(expression):
    """
    Calculate the result of a mathematical expression in Roman numerals.

    Args:
        expression (str): The mathematical expression in Roman numerals.

    Returns:
        str: The result of the calculation in Roman numerals.
    """
    # Implement the logic to parse and calculate the expression
    # This is a placeholder for the actual implementation
    return "XXV"

if __name__ == "__main__":
    expression = " ".join(sys.argv[1:])
    result = calculate(expression)
    print(result)
